p2w-client: Add lib target, make helpers into lib functions there
commit-id:3aeb9ee6
