+++ solana/pyth2wormhole/program/Cargo.toml
@@ -28,3 +28,4 @@ wasm-bindgen = { version = "0.2.74", features = ["serde-serialize"], optional = 
 serde = { version = "1", optional = true}
 serde_derive = { version = "1", optional = true}
 serde_json = { version = "1", optional = true}
+
